1) Working on getting the FSMC to work properly with SRAM ... ran into some problems -- this should be everything once I can prove that a read/write works properly.

+++ Firmware/header/adc.h
@@ -64,7 +64,7 @@ typedef struct
 {
   ADCConfig adcConfig;  // Not using this yet, but ultimately allow apps to decide these params
   uint16 *appSampleBuffer;
-  void  (*appNotifyConversionComplete)(uint8, uint16);
+  void  (*appNotifyConversionComplete)(uint8, uint32);
 } AppADCConfig;
 
 void ADC_init(void);
